+++ addons/yat/src/commands/builtin/Monitor.cs
@@ -13,6 +13,7 @@ namespace YAT.Commands
 	[Option("-fps", null, "Shows the FPS in the performance monitor.", false)]
 	[Option("-os", null, "Shows the OS information in the performance monitor.", false)]
 	[Option("-cpu", null, "Shows the CPU information in the performance monitor.", false)]
+	[Option("-ram", null, "Shows the RAM information in the performance monitor.", false)]
 	public sealed class Monitor : ICommand
 	{
 		public YAT Yat { get; set; }
@@ -47,6 +48,7 @@ namespace YAT.Commands
 			bool fps = (bool)cArgs["-fps"];
 			bool os = (bool)cArgs["-os"];
 			bool cpu = (bool)cArgs["-cpu"];
+			bool ram = (bool)cArgs["-ram"];
 
 			List<Node> components = new();
 			bool instanceValid = GodotObject.IsInstanceValid(_monitorInstance);
@@ -60,6 +62,10 @@ namespace YAT.Commands
 			if (fps) components.Add(GD.Load<PackedScene>(
 				_componentsPath + "fps/Fps.tscn"
 			).Instantiate<Fps>());
+
+			if (ram) components.Add(GD.Load<PackedScene>(
+				_componentsPath + "memory_info/MemoryInfo.tscn"
+			).Instantiate<MemoryInfo>());
 
 			if (os) components.Add(GD.Load<PackedScene>(
 				_componentsPath + "os/Os.tscn"