+++ CHANGELOG.md
@@ -16,6 +16,11 @@ All notable changes to this project will be documented in this file.
 
 -   OverlayOpened & OverlayClosed signals.
 
+### Fixed
+
+-   The terminal toggle button no longer writes to the input
+    (it's a workaround, the input is unfortunately cleared every time the terminal is opened).
+
 ## [1.14.0-beta 2023-12-08]
 
 ### Added